+++ R/DM2.R
@@ -15,7 +15,10 @@
 #' \code{alternative="greater"}, the alternative hypothesis is that method 2 is
 #' more accurate than method 1. For \code{alternative="two.sided"}, the
 #' alternative hypothesis is that method 1 and method 2 have different levels
-#' of accuracy.
+#' of accuracy. The long-run variance estimator can either the
+#' auto-correlation estimator \code{varestimator = "acf"}, or the estimator based
+#' on Bartlett weights \code{varestimator = "bartlett"} which ensures a positive estimate.
+#' Both long-run variance estimators are proposed in Diebold and Mariano (1995).
 #'
 #' @param e1 Forecast errors from method 1.
 #' @param e2 Forecast errors from method 2.
@@ -24,14 +27,18 @@
 #' \code{"less"}.  You can specify just the initial letter.
 #' @param h The forecast horizon used in calculating \code{e1} and \code{e2}.
 #' @param power The power used in the loss function. Usually 1 or 2.
+#' @param varestimator a character string specifying the long-run variance estimator.
+#' Options are \code{"acf"} (default) or \code{"bartlett"}.
 #' @return A list with class \code{"htest"} containing the following
-#' components: \item{statistic}{the value of the DM-statistic.}
-#' \item{parameter}{the forecast horizon and loss function power used in the
-#' test.} \item{alternative}{a character string describing the alternative
-#' hypothesis.} \item{p.value}{the p-value for the test.} \item{method}{a
-#' character string with the value "Diebold-Mariano Test".} \item{data.name}{a
-#' character vector giving the names of the two error series.}
-#' @author George Athanasopoulos
+#' components:
+#' \item{statistic}{the value of the DM-statistic.}
+#' \item{parameter}{the forecast horizon and loss function power used in the test.}
+#' \item{alternative}{a character string describing the alternative hypothesis.}
+#' \item{varestimator}{a character string describing the long-run variance estimator.}
+#' \item{p.value}{the p-value for the test.}
+#' \item{method}{a character string with the value "Diebold-Mariano Test".}
+#' \item{data.name}{a character vector giving the names of the two error series.}
+#' @author George Athanasopoulos and Kirill Kuroptev
 #' @references Diebold, F.X. and Mariano, R.S. (1995) Comparing predictive
 #' accuracy. \emph{Journal of Business and Economic Statistics}, \bold{13},
 #' 253-263.
@@ -47,7 +54,7 @@
 #' f2 <- auto.arima(WWWusage)
 #' accuracy(f1)
 #' accuracy(f2)
-#' dm.test(residuals(f1),residuals(f2),h=1)
+#' dm.test(residuals(f1), residuals(f2), h=1)
 #'
 #' # Test on out-of-sample one-step forecasts
 #' f1 <- ets(WWWusage[1:80])
@@ -56,22 +63,36 @@
 #' f2.out <- Arima(WWWusage[81:100],model=f2)
 #' accuracy(f1.out)
 #' accuracy(f2.out)
-#' dm.test(residuals(f1.out),residuals(f2.out),h=1)
-#'
+#' dm.test(residuals(f1.out), residuals(f2.out), h=1)
 #' @export
-dm.test <- function(e1, e2, alternative = c("two.sided", "less", "greater"), h = 1, power = 2) {
+
+dm.test <- function(e1, e2, alternative = c("two.sided", "less", "greater"), h = 1,
+                    power = 2, varestimator = c("acf", "bartlett")) {
+
   alternative <- match.arg(alternative)
-  d <- c(abs(e1)) ^ power - c(abs(e2)) ^ power
+  varestimator <- match.arg(varestimator)
+
+  d <- c(abs(e1))^power - c(abs(e2))^power
   d.cov <- acf(d, na.action = na.omit, lag.max = h - 1, type = "covariance", plot = FALSE)$acf[, , 1]
-  d.var <- sum(c(d.cov[1], 2 * d.cov[-1])) / length(d)
-  dv <- d.var # max(1e-8,d.var)
+
+  if (varestimator == "acf") {
+    # Original estimator
+    d.var <- sum(c(d.cov[1], 2 * d.cov[-1])) / length(d)
+  } else { # varestimator == "bartlett"
+    # Using Bartlett weights to ensure a positive estimate of long-run-variance
+    m <- h-1
+    b <- seq.int(from = 1,to = m)
+    d.var <- sum(c(d.cov[1], 2*(1-b/(m+1))*d.cov[-1])) / length(d)
+  }
+  dv <- d.var
+
   if (dv > 0) {
     STATISTIC <- mean(d, na.rm = TRUE) / sqrt(dv)
   } else if (h == 1) {
     stop("Variance of DM statistic is zero")
   } else {
-    warning("Variance is negative, using horizon h=1")
-    return(dm.test(e1, e2, alternative, h = 1, power))
+    warning("Variance is negative. Try varestimator = bartlett. Proceeding with horizon h=1.")
+    return(dm.test(e1, e2, alternative, h = 1, power, varestimator))
   }
 
   n <- length(d)
@@ -90,12 +111,13 @@ dm.test <- function(e1, e2, alternative = c("two.sided", "less", "greater"), h =
   structure(
     list(
       statistic = STATISTIC, parameter = PARAMETER,
-      alternative = alternative, p.value = PVAL, method = "Diebold-Mariano Test",
+      alternative = alternative,varestimator=varestimator, p.value = PVAL, method = "Diebold-Mariano Test",
       data.name = c(deparse(substitute(e1)), deparse(substitute(e2)))
     ),
     class = "htest"
   )
 }
+
 
 is.htest <- function(x) {
   inherits(x, "htest")