+++ NEWS.md
@@ -1,3 +1,9 @@
+# forecast 8.17 (25 July 2022)
+  * Updated dm.test() to add alternative variance estimators. (#898)
+  * Added `simulate.tbats()` for simulating from TBATS models.
+  * Added dependency on generics for accuracy() and forecast() (#902)
+  * Bux fixes
+
 # forecast 8.16 (10 January 2022)
   * Fixed `tslm()` incorrectly applying Box-Cox transformations when an `mts`
 is provided to the `data` argument (#886).